[commhistoryd] FsCleanup::onEventDeleted should check if event is actually gone
Anyone can send this D-Bus event and it's generally a good idea to
validate the argument. Customers are complaining that stray delete
events do happen, causing deletion of MMS attachments.

+++ src/fscleanup.cpp
@@ -46,7 +46,12 @@ FsCleanup::FsCleanup(QObject* aParent) :
 void FsCleanup::onEventDeleted(int aEventId)
 {
     DEBUG_("Event" << aEventId << "deleted");
-    deleteFiles(aEventId);
+    CommHistory::DatabaseIO* io = CommHistory::DatabaseIO::instance();
+    if (!io->eventExists(aEventId)) {
+        deleteFiles(aEventId);
+    } else {
+        qWarning() << "Ignoring stray delete event for" << aEventId;
+    }
 }
 
 void FsCleanup::onGroupsDeleted(QList<int> aGroupIds)
